+++ androidx/src/main/java/me/panpf/androidx/app/Fragmentx.java
@@ -18,9 +18,12 @@ package me.panpf.androidx.app;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+
+import org.jetbrains.annotations.NotNull;
 
 import me.panpf.androidx.content.Intentx;
 
@@ -30,7 +33,7 @@ public class Fragmentx {
      * If the own or parent Fragment implements the specified [clazz], it returns its implementation.
      */
     @Nullable
-    public <T> T getImplWithParent(@NonNull Fragment fragment, @NonNull Class<T> clazz) {
+    public static <T> T getImplWithParent(@NonNull Fragment fragment, @NonNull Class<T> clazz) {
         Fragment parent = fragment;
         while (parent != null) {
             if (clazz.isAssignableFrom(parent.getClass())) {
@@ -74,5 +77,31 @@ public class Fragmentx {
         }
 
         return Intentx.safeStartActivity(activity, intent);
+    }
+
+    /**
+     * Instantiate a Fragment and set arguments
+     */
+    public static Fragment instance(@NotNull Class<? extends Fragment> clas, @Nullable Bundle arguments) {
+        Fragment fragment;
+        //noinspection TryWithIdenticalCatches
+        try {
+            fragment = clas.newInstance();
+        } catch (InstantiationException e) {
+            throw new IllegalArgumentException(e);
+        } catch (IllegalAccessException e) {
+            throw new IllegalArgumentException(e);
+        }
+        if (arguments != null) {
+            fragment.setArguments(arguments);
+        }
+        return fragment;
+    }
+
+    /**
+     * Instantiate a Fragment and set arguments
+     */
+    public static Fragment instance(@NotNull Class<? extends Fragment> clas) {
+        return instance(clas, null);
     }
 }